webapp/app/mdeautomator_mcp/requirements-docker.txt
# Requirements for container environments
# This ensures trio and all its dependencies are properly installed

# Core MCP dependencies with trio support
mcp==1.9.4
trio==0.30.0
anyio==4.9.0

# Trio's system dependencies (critical for containers)
attrs>=21.4.0
cffi>=1.15.0
idna>=3.4
outcome>=1.2.0
sniffio>=1.3.0
sortedcontainers>=2.4.0

# Additional system dependencies that trio might need
exceptiongroup>=1.0.4; python_version < "3.11"

# Core application dependencies
pydantic>=2.0.0
httpx>=0.25.0
asyncio-throttle>=1.0.0

# Azure SDK Dependencies
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0

# Additional Dependencies
python-dotenv>=1.0.0
structlog>=23.2.0
tenacity>=8.2.0
typing-extensions>=4.8.0

# Development Dependencies (optional for containers)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.21.0
